I10: honest payment labels; card/manual stays pending (no gateway)

+++ app/models/pos_session.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, Float, Text, func
+from sqlalchemy import Column, Integer, String, DateTime, Numeric, Text, func
 
 from app.core.database import Base
 
@@ -14,6 +14,6 @@ class PosSession(Base):
     customer_name = Column(String(255))
     customer_phone = Column(String(50))
     payment_method = Column(String(50))
-    total = Column(Float, default=0)
+    total = Column(Numeric(12, 2), default=0)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     updated_at = Column(DateTime(timezone=True), server_default=func.now(), onupdate=func.now())
